Ensure API file is closed
`load_text_from_file` did not close the file after opening it.
The `with` statement creates a context which will close the file as if
in a `finally` block.

`get_current_directory` was removed; it is a NOOP, path will be
interpreted as a relative path, unless explicitly written as an
absolute.

N.B.,

It is more pythonic to use `os.getcwd` directly over creating the
wrapper of `get_current_directory`.

Avoid using path separators, they are platform dependent;
use `os.path.join`.