Clean bloat from .git* files
Vast majority of the .gitignore file was not needed, and contains
editor specific ignore files. These belong in a global gitignore, not
project based.

           *nix:   $HOME/.config/git/ignore
        Windows:   %USERPROFILE%\git\ignore

Or custom using
        $ git config --global core.excluedsfile /path/to/global/gitignore

The attributes file contained much unneeded diff information for
irrelevant file formats; perhaps things that belong in a global
attributes file.

As Windows can be used in the development of the project, settings
"text" is relavant. This ensures that all text files are written with LF
line endings, which is the prefered standard.