Add argument parsing for logging
Allow the user to set the verbosity of the logging from the module. If a
file is given, then it will also be used.

The default logging will be DEBUG to catch all. The console will default
to ERROR, to only get important messages. The log file will default to
catching all messages.