Use logger throughout rather than print
Replacing all manual logging via print functions with calls to the
client logger increase portability, integration, and readability.

Initially the client will use the root logger; it will be injected with
the custom logger in the `run` function. It would be an idea to prevent
the global scope if possible.