Fix dictionary based errors
dicts are not guaranteed to be ordered; it just so happens an
implementation detail in CPython 3.6+ results in them being ordered.
This is not true for all implementations of Python compatible with the
3.6 based syntax used. collections.OrderDict is used in place to retain
the order the logging levels.

A collections.ChainMap when updated, edits the first map within it's
structure. To retain semantics of each internal map, adding a blank map
first is a common practice; any updates to the ChainMap will not edit
the other maps, but still keep precedence.