Allow ?lightthebeacons to mention pseudo roles
The user should be able to ping @everyone and those that are @here.

These pseudo roles are not quite returned from querying
`server.roles`.

`server.roles` will return the "role" `@everyone`, with the @, and it's
not considered mentionable. Therefore `everyone` and `here` are added to
be explicitly mentioned using the mantual @-syntax.